requirements.txt
Flask==0.9
Flask-PyMongo==0.1.3
Jinja2==2.6
PIL==1.1.7
Werkzeug==0.8.3
pyenchant==1.6.5
pymongo==2.3
requests==0.14.2
wsgiref==0.1.2
